Indian astrology is divided into following categories
Astrology is a language, if you understand this language, the sky speaks to you. Each and every one of us believe in god and curious to know about our future. In India starting from the birth to death we believe in consulting an astrologer for anything and everything.
Birth Chart 
A birth chart known as janma kundali, kundli, janampatri, Vedic horoscope, janam kundali, Vedic chart, Tewa,  Hindu chart, Teepna etc. )after capturing the positions of one’s stars and planets the future is predicted ..